requirements.txt
Django~=4.2.4
djangorestframework~=3.14.0
psycopg2~=2.9.7
drf-yasg~=1.21.7